+++ webapp/components/survey/ButtonTaxonomyAdd/ButtonTaxonomyAdd.js
@@ -1,26 +1,21 @@
 import React, { useCallback } from 'react'
 import PropTypes from 'prop-types'
-import { matchPath, useHistory, useLocation } from 'react-router'
+
 import axios from 'axios'
 
 import { useI18n } from '@webapp/store/system'
 import { SurveyActions, useSurveyId } from '@webapp/store/survey'
 
-import * as A from '@core/arena'
 import * as Taxonomy from '@core/survey/taxonomy'
 
-import { appModuleUri, designerModules } from '@webapp/app/appModules'
 import { useDispatch } from 'react-redux'
 
 const ButtonTaxonomyAdd = (props) => {
   const i18n = useI18n()
-  const { pathname } = useLocation()
-  const history = useHistory()
+
   const dispatch = useDispatch()
 
-  const inTaxonomiesPath = Boolean(matchPath(pathname, appModuleUri(designerModules.taxonomies)))
-
-  const { onTaxonomyCreated } = props
+  const { onAdd } = props
 
   const surveyId = useSurveyId()
 
@@ -31,11 +26,7 @@ const ButtonTaxonomyAdd = (props) => {
 
     dispatch(SurveyActions.metaUpdated())
 
-    if (inTaxonomiesPath || A.isNull(onTaxonomyCreated)) {
-      history.push(`${appModuleUri(designerModules.taxonomy)}${Taxonomy.getUuid(taxonomyCreated)}`)
-    } else {
-      onTaxonomyCreated(taxonomyCreated)
-    }
+    onAdd(taxonomyCreated)
   }, [])
 
   return (
@@ -47,11 +38,11 @@ const ButtonTaxonomyAdd = (props) => {
 }
 
 ButtonTaxonomyAdd.propTypes = {
-  onTaxonomyCreated: PropTypes.func,
+  onAdd: PropTypes.func,
 }
 
 ButtonTaxonomyAdd.defaultProps = {
-  onTaxonomyCreated: () => ({}),
+  onAdd: () => ({}),
 }
 
 export default ButtonTaxonomyAdd
